week7/requirements.txt
Flask
pymongo
flasgger
jsonschema==4.17.3
python-dotenv